requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
yfinance==0.2.50
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
scikit-learn==1.5.2
joblib==1.4.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
apscheduler==3.10.4
requests==2.32.3
openpyxl==3.1.5
pytz==2024.2
